Resolved get nonexistent thread 500 bug + rewrote tests to test class per function
